rebase: class WC_woocommerce_enviosimples
